chore: move attestation verification code to nilcc-verifier
