+++ docker/nilcc-attester.dockerfile
@@ -1,7 +1,7 @@
 FROM rust:1.86-alpine AS build
 
 WORKDIR /opt/nillion
-RUN apk add --no-cache musl-dev git perl make
+RUN apk add --no-cache musl-dev git
 
 COPY . .
 RUN cargo build --release --locked -p nilcc-attester
