+++ nilcc-attester/Cargo.toml
@@ -10,16 +10,10 @@ axum = "0.8"
 clap = { version = "4.5", features = ["derive"] }
 config = { version = "0.15", default-features = false, features = ["yaml"] }
 hex = { version = "0.4", features = ["serde"] }
-openssl = { version = "^0.10", features = ["vendored"] }
 rand = "0.9"
-reqwest = "0.12"
 serde = { version = "1.0", features = ["derive"] }
-sev = { version = "6.1", default-features = false, features = ["openssl", "snp"] }
+sev = { version = "6.1", default-features = false, features = ["snp"] }
 thiserror = "2.0"
 tokio = { version = "1.45", features = ["rt-multi-thread", "macros", "signal"] }
 tracing = "0.1"
 tracing-subscriber = "0.3"
-x509-parser = { version = "0.17.0", features = ["verify"] } 
-
-[dev-dependencies]
-rstest = { version = "0.25", default-features = false }
