+++ nilcc-attester/src/lib.rs
@@ -1,5 +1,3 @@
-pub mod certs;
 pub mod config;
 pub mod report;
 pub mod routes;
-pub mod verify;
