+++ nilcc-attester/src/main.rs
@@ -1,11 +1,5 @@
 use clap::Parser;
-use nilcc_attester::{
-    certs::{CertFetchPolicy, DefaultCertificateFetcher},
-    config::Config,
-    report::request_hardware_report,
-    routes::build_router,
-    verify::ReportVerifier,
-};
+use nilcc_attester::{config::Config, routes::build_router};
 use tokio::{net::TcpListener, signal};
 use tracing::{error, info};
 
@@ -47,21 +41,6 @@ async fn main() {
             std::process::exit(1);
         }
     };
-
-    if config.attestation.verify_on_startup {
-        info!("Getting report to use during initial verification");
-        let report = request_hardware_report(rand::random()).expect("failed to get hardware report");
-
-        info!("Verifying report");
-        let fetcher = DefaultCertificateFetcher::new(CertFetchPolicy::PreferHardware);
-        let mut verifier = ReportVerifier::new(Box::new(fetcher));
-        if let Some(processor) = config.attestation.processor {
-            info!("Forcing processor to be {processor:?}");
-            verifier = verifier.with_processor(processor);
-        }
-        verifier.verify_report(report).await.expect("failed to fetch certs");
-        info!("Verification succeeded");
-    }
 
     let bind_endpoint = &config.server.bind_endpoint;
     info!("Running server on {bind_endpoint}");
