+++ nilcc-attester/src/routes/health.rs
@@ -1,5 +1,7 @@
-use axum::response::{IntoResponse, Response};
-use reqwest::StatusCode;
+use axum::{
+    http::StatusCode,
+    response::{IntoResponse, Response},
+};
 
 pub(crate) async fn handler() -> Response {
     (StatusCode::OK, "OK").into_response()
