+++ nilcc-verifier/Cargo.toml
@@ -4,11 +4,21 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
+async-trait = "0.1"
 anyhow = "1.0"
 nilcc-attester = { path = "../nilcc-attester" }
 clap = { version = "4.5", features = ["derive"] }
+hex = "0.4"
+openssl = { version = "^0.10", features = ["vendored"] }
 reqwest = "0.12"
+serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
+sev = { version = "6.1", default-features = false, features = ["openssl", "snp"] }
+thiserror = "2.0"
 tokio = { version = "1.45", features = ["rt-multi-thread", "macros"] }
 tracing = "0.1"
 tracing-subscriber = "0.3"
+x509-parser = { version = "0.17.0", features = ["verify"] } 
+
+[dev-dependencies]
+rstest = { version = "0.25", default-features = false }
